fix: ArticleWrite 내보내기 방식 수정
